chore(mypy): burn 32 -> 0 and drop MYPY_MAX cap (#18)
Systematic cleanup of the 32 mypy errors that were being carried as a
regression cap. CI now requires zero mypy errors; future type errors
must be fixed in the same PR rather than raising the cap.

Fixes by cluster:

1. ReagentProfile vs CrosslinkerProfile confusion (9 errors in
   protocols/mechanism_data.py + module2_functionalization/
   modification_steps.py). Root cause: both branches of a try/except
   ladder assigned to the same 'profile' variable with different
   types. Fixed by renaming the M2 branch variable to 'm2_profile'
   and narrowing 'k_ref' access via a local alias.

2. Family context Union in visualization/tabs/tab_m1.py (6 errors).
   'fctx' was assigned AlginateContext in the first branch and later
   reassigned to CelluloseContext / PLGAContext in the dispatch
   ladder. Fixed by annotating 'fctx: Union[AlginateContext,
   CelluloseContext, PLGAContext]' and wrapping each family-specific
   attribute access in an isinstance() guard.

3. Optional-None defaults (3 errors in __init__.py, pipeline/
   orchestrator.py, level1_emulsification/energy.py). Changed
   'x: T = None' to 'x: T | None = None' (or 'Optional[T]') and
   added None-narrowing assertions.

4. numpy / object narrowing (4 errors in level1/validation.py,
   level2/pore_analysis.py, digital_twin/enkf.py). Cast numpy scalars
   to Python floats; normalised rng.normal() output with
   np.asarray(); wrapped numpy bool with bool().

5. CahnHilliard subclass assignment (3 errors in level2_gelation/
   solver.py). Renamed the per-branch 'solver' variable to
   'solver_ternary' and 'solver_1d' so each branch's type does not
   leak into the next branch.

6. Scalar/array unions in module3_performance/catalysis/packed_bed.py
   (2 errors). Cast first_order_deactivation return to float at
   the call site; normalised activity_history to ndarray via
   np.asarray.

7. Misc one-offs: datatypes.py:922 str() coercion on tier_value;
   pipeline/orchestrator.py:364 filter None from _manifests before
   passing to RunReport; tab_m1.py:372 600.0 -> 600 to match the
   slider's int type.

CI update:
  .github/workflows/ci.yml: MYPY_MAX 32 -> 0; rename step label;
  update comment to document the burndown history.

CLAUDE.md: updated 'CI gates' section to reflect the new zero-errors
policy.

No runtime, API, or test change. Infra-only cleanup. All 924 pytest
tests still pass.

+++ src/emulsim/level2_gelation/pore_analysis.py
@@ -342,8 +342,8 @@ def morphology_descriptors(phi_field: np.ndarray, grid_spacing: float,
         if chord > 0:
             v_chords.append(chord * grid_spacing)
 
-    mean_h = np.mean(h_chords) if h_chords else 1.0
-    mean_v = np.mean(v_chords) if v_chords else 1.0
+    mean_h = float(np.mean(h_chords)) if h_chords else 1.0
+    mean_v = float(np.mean(v_chords)) if v_chords else 1.0
     anisotropy = abs(mean_h - mean_v) / max(mean_h, mean_v) if max(mean_h, mean_v) > 0 else 0.0
 
     # 3. Connectivity: simple flood fill from top-left pore pixel
